Add methods `.as_raw_str/slice` to shared string character types and Integer
This methods helps avoiding lifetimes problems for crate users, when building an
ephemeral object from input and returning data from this object (destroying it).
This should work using `AsRef`, however the `AsRef` trait definition does not define the
lifetimes as expected and does not allow to return a reference that can outlive
the object.